Limited max number of onto-neighbors considered for topsim
KLUDGE: for some classes such as 'protein' we might get very large
neighborhoods (80k+ nodes, 400k+ for 'gene'), so for now we just
randomly pick a subset, however, something better needs to be done -
e.g., picking a subset based on pagerank, for example.  In a test to
pick the right threshold, we found that 99.4% of WD top-200k classes
(by pagerank) have <= 10,000 onto-neighbors.